Revert "added candidate single (view resume page)"
