+++ mongodb/controllers/Applicant.js
@@ -15,18 +15,7 @@ exports.Applicant_create = function (req, res) {
             name: req.body.name,
             email: req.body.email,
             password: req.body.password,
-            phoneno: req.body.phoneno,
-            gender: "",
-            dob: new Date(),
-            qualification: "",
-            experience: "",
-            currentJob: "",
-            currentSalary: "",
-            currentCompany: "",
-            about: "",
-            image: "https://img.flaticon.com/icons/png/512/149/149071.png?size=1200x630f&pad=10,10,10,10&ext=png&bg=FFFFFFFF",
-            address: {city: "", state: "", country: "", pincode: ""},
-            socialMedia: {facebook: "", linkedin: "", twitter: "", github: ""}
+            phoneno: req.body.phoneno
           });
           //Hash password before saving in database
           bcrypt.genSalt(10, (err, salt) => {
@@ -62,12 +51,7 @@ exports.Applicant_login = function (req, res) {
                     username: applicant.username,
                     name: applicant.name,
                     email: applicant.email,
-                    phoneno: applicant.phoneno,
-                    image: applicant.image,
-                    Job: applicant.currentJob,
-                    Company: applicant.currentCompany,
-                    City: applicant.address.city,
-                    Country: applicant.address.country
+                    phoneno: applicant.phoneno
                 };
                 jwt.sign(
                     payload,
@@ -113,7 +97,7 @@ exports.Applicant_details = function (req, res) {
 exports.Applicant_delete = function (req, res) {
     if(req.params.field=="edu"){
         Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$pull: {"resume.education": {etype: req.body.etype}}}).then((applicant) => {
-            res.redirect("http://localhost:3000/buildResume");
+            res.redirect("http://localhost:3000/resume");
         })
         .catch((error) => {
             console.log('Applicant Education Details Not Deleted! ', error);
@@ -126,7 +110,7 @@ exports.Applicant_delete = function (req, res) {
             operator['$set'] = selector;
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, operator, {$new: true}).then((applicant) => {
                     Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$pull:{"resume.experience": {company: "delete"}}}).then((applicant) => {
-                        res.redirect("http://localhost:3000/buildResume#experience");
+                        res.redirect("http://localhost:3000/resume#experience");
                     })
                     .catch((error) => {
                         console.log('Work Experience Not Deleted! ', error);
@@ -143,7 +127,7 @@ exports.Applicant_delete = function (req, res) {
         operator['$set'] = selector;
             Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, operator, {$new: true}).then((applicant) => {
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$pull:{"resume.projects": {title: "delete"}}}).then((applicant) => {
-                    res.redirect("http://localhost:3000/buildResume#projects");
+                    res.redirect("http://localhost:3000/resume#projects");
                 })
                 .catch((error) => {
                     console.log('Project Not Deleted! ', error);
@@ -160,7 +144,7 @@ exports.Applicant_delete = function (req, res) {
         operator['$set'] = selector;
             Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, operator, {$new: true}).then((applicant) => {
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$pull:{"resume.achievements": {atitle: "delete"}}}).then((applicant) => {
-                    res.redirect("http://localhost:3000/buildRresume#projects");
+                    res.redirect("http://localhost:3000/resume#projects");
                 })
                 .catch((error) => {
                     console.log('Achievement Not Deleted! ', error);
@@ -172,7 +156,7 @@ exports.Applicant_delete = function (req, res) {
     }
     else if(req.params.field=="skill"){
         Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$pull:{"resume.skills": req.body.skill}}).then((applicant) => {
-                    res.redirect("http://localhost:3000/buildResume#projects");
+                    res.redirect("http://localhost:3000/resume#projects");
                 })
                 .catch((error) => {
                     console.log('Skill Not Deleted! ', error);
@@ -181,38 +165,16 @@ exports.Applicant_delete = function (req, res) {
 };
 
 exports.Applicant_update = function (req, res) {
-    if(req.params.field=="personal"){
-        Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, req.body, {$new: true}).then((applicant) => {
-            if(applicant){
-                res.redirect("http://localhost:3000/profile#personal");
-            } //update
-        })
-        .catch((error) => {
-            console.log('Applicant Not found! ', error);
-        });
-    }
-    if(req.params.field=="categories")
-    {
-        Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$set: {"categories": req.body.categories}}, {$new: true}).then((applicant) =>{
-                if(applicant)
-                {
-                    return res.json(applicant);
-                }
-            })
-            .catch((error) => {
-                console.log('Applicant Not found! ', error);
-            });
-    }
     if(req.params.field=="edu"){
             Applicant.findOneAndUpdate({$and: [{Applicant_Id: req.params.aid}, {"resume.education.etype": req.body.etype}]}, {$set: {"resume.education.$": req.body}}, {$new: true}).then((applicant) => {
                 //check if edu details exist based on etype -> if exists -> update else push
                 if(applicant){
-                    res.redirect("http://localhost:3000/buildResume#res");
+                    res.redirect("http://localhost:3000/resume#res");
                 } //update
                 else{
                     Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$push: {"resume.education": req.body}}, function (err, data) {
                         if (err) console.log(err);
-                        res.redirect("http://localhost:3000/buildResume#res");
+                        res.redirect("http://localhost:3000/resume#res");
                     });
                 } //push
             })
@@ -229,7 +191,7 @@ exports.Applicant_update = function (req, res) {
             {
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$push: {"resume.experience": req.body}}, function (err, data) {
                     if (err) console.log(err);
-                    res.redirect("http://localhost:3000/buildResume#experience");
+                    res.redirect("http://localhost:3000/resume#experience");
                 });
             }
             else{ //update
@@ -238,7 +200,7 @@ exports.Applicant_update = function (req, res) {
                 operator['$set'] = selector;
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, operator, {$new: true}).then((applicant) => {
                     //check if edu details exist based on etype -> if exists -> update else push
-                res.redirect("http://localhost:3000/buildResume#experience");
+                res.redirect("http://localhost:3000/resume#experience");
                 })
                 .catch((error) => {
                     console.log('Applicant Not found! ', error);
@@ -252,7 +214,7 @@ exports.Applicant_update = function (req, res) {
             {
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$push: {"resume.projects": req.body}}, function (err, data) {
                     if (err) console.log(err);
-                    res.redirect("http://localhost:3000/buildResume#projects");
+                    res.redirect("http://localhost:3000/resume#projects");
                 });
             }
             else{ //update
@@ -261,7 +223,7 @@ exports.Applicant_update = function (req, res) {
                 operator['$set'] = selector;
                 Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, operator, {$new: true}).then((applicant) => {
                     //check if edu details exist based on etype -> if exists -> update else push
-                res.redirect("http://localhost:3000/buildResume#projects");
+                res.redirect("http://localhost:3000/resume#projects");
                 })
                 .catch((error) => {
                     console.log('Applicant Not found! ', error);
@@ -275,7 +237,7 @@ exports.Applicant_update = function (req, res) {
                 {
                     Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$push: {"resume.achievements": req.body}}, function (err, data) {
                         if (err) console.log(err);
-                        res.redirect("http://localhost:3000/buildResume#achievements");
+                        res.redirect("http://localhost:3000/resume#achievements");
                     });
                 }
                 else{ //update
@@ -284,7 +246,7 @@ exports.Applicant_update = function (req, res) {
                     operator['$set'] = selector;
                     Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, operator, {$new: true}).then((applicant) => {
                         //check if edu details exist based on etype -> if exists -> update else push
-                    res.redirect("http://localhost:3000/buildResume#achievements");
+                    res.redirect("http://localhost:3000/resume#achievements");
                     })
                     .catch((error) => {
                         console.log('Applicant Not found! ', error);
@@ -293,7 +255,7 @@ exports.Applicant_update = function (req, res) {
             }
         else if(req.params.field=="skill"){
             Applicant.findOneAndUpdate({Applicant_Id: req.params.aid}, {$push: {"resume.skills": {$each: req.body.skills}}}).then((data) =>{
-                res.redirect("http://localhost:3000/buildResume#skills");
+                res.redirect("http://localhost:3000/resume#skills");
             })
             .catch((error) => {
                 console.log('Applicant Not found! ', error);
