+++ mongodb/package.json
@@ -11,7 +11,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@vonage/server-sdk": "^2.10.8",
     "alert": "^5.0.10",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
@@ -22,8 +21,6 @@
     "is-empty": "^1.2.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.12.6",
-    "nexmo": "^2.9.1",
-    "nodemailer": "^6.6.0",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
     "sift": "^13.5.3",
