+++ package.json
@@ -21,7 +21,6 @@
     "react-countup": "^4.3.3",
     "react-dom": "^17.0.1",
     "react-icons": "^4.2.0",
-    "react-images-upload": "^1.2.8",
     "react-popper": "^2.2.4",
     "react-redux": "^7.2.4",
     "react-responsive-modal": "^6.0.1",
@@ -29,7 +28,6 @@
     "react-scripts": "4.0.2",
     "react-select": "^4.3.0",
     "react-slick": "^0.28.1",
-    "react-sticky-header": "^0.2.0",
     "react-visibility-sensor": "^5.1.1",
     "reactstrap": "^8.9.0",
     "reactstrap-date-picker": "0.0.11",
