+++ src/Components/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
 import Home from './Home/Home';
+import Dummy from './Dummy/dummy';
 import Dashboard from './Applicant/Dashboard';
-import BuildResume from './Applicant/BuildResume';
 import Resume from './Applicant/Resume';
 import jwt_decode from "jwt-decode";
 import setAuthToken from "../utils/setAuthToken";
@@ -39,7 +39,6 @@ function App(){
 			<Router>
 				<Switch>
 					<Route path="/" component={Home} exact/>
-					<PrivateRoute path="/buildResume" component={BuildResume} exact/>
 					<PrivateRoute path="/resume" component={Resume} exact/>
 					<PrivateRoute path="/Dashboard" component={Dashboard} exact/>
 				</Switch>
