+++ src/Components/Applicant/Dashboard.js
@@ -26,7 +26,7 @@ class Dashboard extends Component
     
     return(
         <div>
-            <Hero logout={this.onLogoutClick} user={user}/>
+            <Hero logout={this.onLogoutClick} name={user.name}/>
             <div className="row mx-0 mt-5 pb-5">
                 <div className="col-lg-3" style={{borderRight: "1px solid #eee"}}>
                     <SideNav logout={this.onLogoutClick} />
