+++ src/Components/Applicant/Resume.js
@@ -1,107 +1,151 @@
-import React, {useEffect, useState} from 'react';
-import { Col, Row, Form, FormGroup, Label, Input } from 'reactstrap';
-import {SiMinutemailer} from 'react-icons/si';
-import { connect } from "react-redux";
-import Details from './Details';
+import React, { useState }  from 'react';
+import Hero from './Hero';
+import SideNav from './SideNav';
 import Footer from '../Home/Footer';
-import { Navbar, NavItem, Nav,NavLink } from 'reactstrap';
-import '../Styles/Applicant/Resume.css'
+import CircularProgress from './CircularProgress';
+import EducationForm from './EducationForm';
+import ExperienceForm from './ExperienceForm';
+import ProjectForm from './ProjectForm';
+import AchievementForm from './AchievementForm';
+import Skills from './SkillsForm';
+import Details from './Details';
+import "../Styles/Applicant/Resume.css";
+import {AiOutlineProfile, AiOutlineTrophy} from 'react-icons/ai';
+import {IoIosAddCircle} from 'react-icons/io';
+import {CgWorkAlt} from 'react-icons/cg';
+import {BsListCheck, BsGraphUp} from 'react-icons/bs';
+import { logoutUser } from "../../actions/authActions";
+import { connect } from "react-redux";
+
 
 
 function Resume(props){
-    const { user } = props.auth;
-    // const url = "http://localhost:1234/Mail/"+user.email;
-	const [sticky, setIsSticky] = useState(false);
+	const onLogoutClick = e => {
+        e.preventDefault();
+        props.logoutUser();
+     };
+     const { user } = props.auth;
+    
+    
+    const [addEdu, setAddEdu] = useState(false);
+	const [addExp, setAddExp] = useState(false);
+	const [addSkill, setAddSkill] = useState(false);
+    const [addProject, setAddProject] = useState(false);
+    const [addAchievement, setAddAchievement] = useState(false);
 
-	
-	const handleSticky = () => {
-		if(window.scrollY >= 500)
-		{
-		   setIsSticky(true);
-		}
-		else
-		{
-		   setIsSticky(false);
-		}
+    const changeAddEdu = () =>{
+		setAddEdu(true);
+	}
+	const changeAddExp = () =>{
+		setAddExp(true);
+	}
+    const changeAddSkill = () =>{
+		setAddSkill(true);
+	}
+    const changeAddProject = () =>{
+		setAddProject(true);
+	}
+    const changeAddAchievement = () =>{
+		setAddAchievement(true);
 	}
 
-	useEffect(()=>{
-		window.addEventListener('scroll', handleSticky);
-		return () => { window.removeEventListener('scroll', () => handleSticky); }
-	});
-
-	const url = "http://localhost:1234/sms/"+user.phoneno;
-
-	
-
-	return ( <div className="viewResume">
-		<div className="header row mx-0"></div>
-		<div className="row mx-0 mt-5 candDetails">
-			<div className="col-md-4 socialMediaLinks"><Details aid={user.aid} field={"socialMedia"} edit={false}/></div>
-			<div className="col-md-4 personalDetails"><Details aid={user.aid} field={"personalDetails"} edit={false}/></div>
-			<div className="col-md-4 contact"><a href="#contact"><span className="r-icons"><SiMinutemailer size={40}/></span> Contact {user.name.split(" ")[0]}</a></div>
-		</div>
-		<div> 
-			<div className={sticky ? "resnav active" : "resnav"}>
-                <ul className="mr-auto ml-auto pt-2 pb-2 pl-0 pr-0 resume_nav">
-					<li><a className="reslink" href="#about">About</a></li>
-					<li><a className="reslink" href="#edu">Education</a></li>
-					<li><a className="reslink" href="#exp">Work Experience</a></li>
-					<li><a className="reslink" href="#skill">Skills</a></li>
-					<li><a className="reslink" href="#proj">Portfolio</a></li>
-					<li><a className="reslink" href="#ach">Achievements</a></li>
-            	</ul>
+	return(
+	<div>
+        <Hero logout={onLogoutClick} name={user.name}/>
+        <div className="row mx-0 mt-5 pb-5">
+            <div className="col-lg-3" style={{borderRight: "1px solid #eee"}}>
+                <SideNav logout={onLogoutClick} />
+                <CircularProgress/>
             </div>
-		</div>
-	 	<Row className="mx-0 rescontent">
-	 		<Col sm={8} className="pl-3 pr-2 leftside">
-				<div id="about"><Details aid={user.aid} field={"about"} edit={false}/></div>
-				<div id="edu"><Details aid={user.aid} field={"edu"} edit={false}/></div>
-				<div id="exp"><Details aid={user.aid} field={"exp"} edit={false}/></div>
-				<div id="proj"><Details aid={user.aid} field={"proj"} edit={false}/></div>
-				<div id="ach"><Details aid={user.aid} field={"ach"} edit={false}/></div>
-			</Col>
-	 		<Col sm={4} className="pl-0 rightside">
-			 	<div id="skill"><Details aid={user.aid} field={"skill"} edit={false}/></div>
-				<div id="overview"><Details aid={user.aid} field={"overview"} edit={false}/></div>
-				<div id="contact">
-					<div className="row mt-5 pt-2 mr-0 field">
-						<h4>Contact</h4>
-					</div>
-					<div className="row showMail">
-						<div className="col-md-12">
-						<Form className="pt-3" action={url} method="POST">
-							<FormGroup row className="px-3">
-								<Input type="text" name="name" placeholder="Name" required={true}></Input>
-							</FormGroup>
-							<FormGroup row className="px-3">
-								<Input type="tel" name="contact" placeholder="Contact No." required={true}></Input>
-							</FormGroup>
-							<FormGroup row className="px-3">
-								<Input type="email" name="email" placeholder="Email (optional)"></Input>
-							</FormGroup>
-							<FormGroup row className="px-3">
-								<Input type="text" name="subject" placeholder="Subject (optional)" required={true}></Input>
-							</FormGroup>
-							<FormGroup row className="px-3">
-								<Input type="textarea" name="msg" maxlength="100" placeholder="Message (not more than 100 characters)" required={true}></Input>
-							</FormGroup>
-							<FormGroup row className="px-5">
-								<Input type="submit" value="Send Message" className="Submitbutton"></Input>
-							</FormGroup>
-						</Form>
-						</div>
-					</div>
-				 </div>
-			</Col>
-		</Row>
-		<div>
-				 
-	 			
-		</div>
-		<div className="row mb-5"></div>
-		<Footer/>
-	</div>);
+            <div className="col-lg-9">
+                <div className="resume-container" id="res">
+        
+                    <div className="resume-head row">
+                        <div className="col md-4">
+                        <h4>Background <span className="r-icons"><AiOutlineProfile size={30}/></span></h4> 
+                        </div>
+                        <div className="col md-5"></div>
+                        <div className="col-md-3 addEdu">
+                            <button onClick={changeAddEdu}><span className="r-icons"><IoIosAddCircle size={30}/></span> Add Education</button>
+                        </div>
+                        <hr className="resume"></hr>
+                    </div>
+            
+                    <div id="education">
+                        <Details aid={user.aid} field={"edu"}/>
+                        {addEdu && <EducationForm aid={user.aid} update={false}/>}
+                    </div>
+		    
+                    <div className="experience-head row mt-5">
+                        <div className="col md-4">
+                            <h4>Work experience <span className="r-icons"><CgWorkAlt size={30}/></span></h4>
+                        </div>
+                        <div className="col md-5"></div>
+                        <div className="col-md-3 addExp">
+                            <button onClick={changeAddExp}><span className="r-icons"><IoIosAddCircle size={30}/></span> Add Experience</button>
+                        </div>
+                        <hr className="resume"></hr>
+                    </div>
+
+                    <div id="experience">
+                        <Details aid={user.aid} field={"exp"}/>
+                        {addExp && <ExperienceForm aid={user.aid} update={false}/>}
+                    </div>
+
+                    <div className="skill-head row mt-5">
+                        <div className="col md-4">
+                            <h4>Skills <span className="r-icons"><BsListCheck size={30}/></span></h4>
+                        </div>
+                        <div className="col md-5"></div>
+                        <div className="col-md-3 addSkill">
+                            <button onClick={changeAddSkill}><span className="r-icons"><IoIosAddCircle size={30}/></span> Add Skill</button>
+                        </div>
+                        <hr className="resume"></hr>
+                    </div>
+
+                    <div id="skills">
+                        <Details aid={user.aid} field={"skill"}/>
+                        {addSkill && <Skills aid={user.aid}/>}
+                    </div>
+
+                    <div className="project-head row mt-5">
+                        <div className="col md-4">
+                            <h4>Portfolio <span className="r-icons"><BsGraphUp size={30}/></span></h4>
+                        </div>
+                        <div className="col md-5"></div>
+                        <div className="col-md-3 addProject">
+                            <button onClick={changeAddProject}><span className="r-icons"><IoIosAddCircle size={30}/></span> Add Project</button>
+                        </div>
+                        <hr className="resume"></hr>
+                    </div>
+
+                    <div id="projects">
+                        <Details aid={user.aid} field={"proj"}/>
+                        {addProject && <ProjectForm aid={user.aid} update={false}/>}
+                    </div>
+
+                    <div className="achievement-head row mt-5">
+                        <div className="col md-4">
+                            <h4>Achievements<span className="r-icons"><AiOutlineTrophy size={30}/></span></h4>
+                        </div>
+                        <div className="col md-5"></div>
+                        <div className="col-md-3 addAchievement">
+                            <button onClick={changeAddAchievement}><span className="r-icons"><IoIosAddCircle size={30}/></span> Add Achievement</button>
+                        </div>
+                        <hr className="resume"></hr>
+                    </div>
+
+                    <div id="achievements">
+                        <Details aid={user.aid} field={"ach"}/>
+                        {addAchievement && <AchievementForm aid={user.aid} update={false}/>}
+                    </div>
+
+                </div>
+            </div>
+        </div>
+        <Footer/>
+    </div>
+	);
 }
 
 const mapStateToProps = state => ({
@@ -109,5 +153,6 @@ const mapStateToProps = state => ({
 });
 export default 
 connect(
-	mapStateToProps
+	mapStateToProps,
+	{ logoutUser }
 )(Resume);
