+++ src/Components/Styles/Applicant/RenderEdu.css
@@ -1,10 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Quicksand&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600&display=swap');
 
-.details{
-	font-family: 'Quicksand';
-}
-
 .details > hr{
  margin-left: 0%;
  width: 85%;
@@ -14,16 +10,11 @@
 	color: #444;
 }
 
-.etype > h5{
-	font-weight: bold;
-}
-
 .ins{
 	margin-top: 0.5rem;
     width: 100%;
     font-size: 16px;
-	font-weight: bold;
-	color: #000;
+	color: #001;
 }
 
 
@@ -35,11 +26,11 @@
 }
 
 .details > ul > li {
-	color: #000 !important;
+	color: #001 !important;
 	margin-top: 0.5rem !important;
 }
 
 .details > ul> li:hover{
-	color: #000 !important;
+	color: #001 !important;
 }
