+++ src/Components/Styles/Applicant/Resume.css
@@ -2,206 +2,23 @@
 @import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600&display=swap');
 
 
-.viewResume{
-	font-family: 'Quicksand';
-}
 
-.view > hr{
+.resume {
+	margin-left: 1%;
 	margin-top: 0%;
-	width: 80%;
-	margin-left: 4%;
+	width: 95%;
 }
 
-.field > h3{
-	font-weight: bold;
- }
- 
-.view > h4{
-	text-align: center;
-	font-weight: bold;
- }
- 
-.currentJob{
-	font-weight: bold;
+.r-icons{
+	color: #e9896a;
+}
+
+
+.resume-container button{
+	background-color: transparent !important;
+}
+.resume-container button:hover {
 	color: #e9896a !important;
 }
 
-.currentCompany{
-	margin-left: 0.5rem;
-	font-weight: bold;
-	color: black;
-}
-
-.view > ul > li{
-	color: #777;
-	font-weight: bold;
-}
-
-.dob{
-	margin-bottom: 1%;
-	margin-right: 0.5rem;
-}
-
-.address{
-	margin-right: 0.5rem;
-}
-
-.about > p{
-	margin-left: 2%;
-}
-
-.header{
-	font-family: 'Quicksand';
-	background: linear-gradient(0deg, rgba(133, 128, 125, 0.7), rgba(20,20,20, 0.7)), url("https://images.unsplash.com/photo-1515378791036-0648a3ef77b2?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=750&q=80") no-repeat;
-	background-size: cover;
-	margin-top: -500px;
-	height: 800px;
-}
-
-.profileImg{
-	position: relative;
-	padding: 1%;
-	border-radius: 50%;
-    height: 180px;
-	width: 180px;
-    border: 3px solid white;
-	margin-top: -180px;
-}
-
-.candDetails{
-	margin-top: 4%;
-}
-
-
-.resnav{
-	margin: 1% 3%;
-	height: 50px;
-	border: 1px solid #aaa;
-    border-radius: 10px;
-	background-color: white !important;
-}
-
-.resnav.active{
-	position: fixed;
-	height: 70px;
-	width: 100%;
-	top:0;
-	left: 0;
-	margin: 0%;
-	padding-top: 1%;
-    z-index: 999;
-}
-
-.rescontent{
-	z-index: -1;
-}
-.resnav > ul{
-	text-align: center;
-}
-
-.resnav > ul > li{
-	display: inline-block;
-	position: relative;
-	margin: 0% 2%;
-	font-size: 18px;
-	font-weight: bold;
-	color: #777;
-}
-
-.reslink:hover{
-	color: #000 !important;
-}
-
-.resnav > ul > li >a::before {
-	transition: 200ms;
-	height: 2px;
-	content: "";
-	position: absolute;
-	background-color: #e9896a;
-	width: 0%;
-	bottom: 1px;
-  }
-  
-.reslink:hover::before {
-	width: 100%;
-}
-
-.personalDetails{
-	text-align: center;
-}
-
-.overview{
-	margin-right: 10% !important;
-	border-radius: 10px;
-	border: 1px solid #aaa;
-}
-
-.overview > ul{
-	padding-bottom: 10px;
-	padding-left: 10px;
-}
-
-.overview > ul > li{
-	color: #000 !important;
-	font-size: 15px;
-}
-
-.overview > ul > li > .r-icons{
-	margin-right: 1rem;
-}
-
-
-.overview_value{
-	margin-left: 0.5rem;
-	color: #000;
-	font-size: 14px;
-	font-style: italic;
-	font-weight: bold;
-}
-
-.socialMedia > li{
-	display: inline-block;
-	margin: 2%;
-}
-
-.socialMedia > li > a{
-	text-decoration: none !important;
-}
-
-.socialMedia > li > a:hover{
-	text-decoration: none !important;
-}
-
-.socialMediaLinks {
-	padding-top: 5%;
-	text-align: center;
-}
-
-.contact{
-	text-align: center;
-	padding-top: 5%;
-}
-.contact > a{
-	font-weight: bold;
-	font-size: 20px;
-	color: black !important;
-	text-decoration: none !important;
-}
-
-.contact > a:hover{
-	font-weight: bold;
-	font-size: 20px;
-	color: #e9896a !important;
-	text-decoration: none !important;
-}
-
-.candname{
-	margin-top: 2%;
-}
-
-.showMail{
-	margin-right: 10% !important;
-	border-radius: 10px;
-	border: 1px solid #aaa;
-}
