+++ src/Components/Styles/Home/Footer.css
@@ -50,7 +50,7 @@
 
 .Footer
 {
-    height: 18rem;
+    height: 21rem;
     background-color: #111;
 }
